refactor: useAddFromBasket
Added hook `useAddFromBasket` which replaces the previous stripes-connect query manifest entry in AgreementCreateRoute/AgreementEditRoute.

However this is NOT a full replacement for the basket system, that still relies on stripes-connect caching for state management across the app
